# Distributor: Refactor OTLP handler