#### What this PR does
Refactor distributor OTLP handler to use `util.ParseProtoReader`, to consolidate with the normal push handler.

There's also some optimization of the OTLP handler in the bargain:

```
✗ benchstat otlp-proto.txt otlp-proto-refactor.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/distributor
cpu: AMD Ryzen 9 3950X 16-Core Processor            
                        │ otlp-proto.txt │      otlp-proto-refactor.txt      │
                        │     sec/op     │   sec/op     vs base              │
OTLPHandler/protobuf-32     1039.2µ ± 2%   960.3µ ± 2%  -7.59% (p=0.002 n=6)

                        │ otlp-proto.txt │       otlp-proto-refactor.txt       │
                        │      B/op      │     B/op      vs base               │
OTLPHandler/protobuf-32     532.3Ki ± 0%   471.6Ki ± 0%  -11.41% (p=0.002 n=6)

                        │ otlp-proto.txt │      otlp-proto-refactor.txt      │
                        │   allocs/op    │  allocs/op   vs base              │
OTLPHandler/protobuf-32      7.141k ± 0%   7.135k ± 0%  -0.08% (p=0.002 n=6)
```

TODO:

- [x] Test in dev

#### Which issue(s) this PR fixes or relates to


#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [ ] [`about-versioning.md`](/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.